Implementation of cart item delete
Created delete cart item in cartservice and implemented the method on the cart page.
Added html delete button
Added delete button css
ticket GUID and date were not matching on delete, reason new guid and date were been created when foreach iteration in the cart page was done.
set ticket guid and date created to be set when ticket is added to cart